backend/requirements.txt
# 🚀 Backend Requirements for AllerScan API
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication & Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Machine Learning & Data Processing
scikit-learn>=1.4.0
joblib>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# API & Web
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-multipart==0.0.6

# Database - MySQL
PyMySQL==1.1.0
SQLAlchemy>=2.0.36
mysql-connector-python==8.2.0

# Environment & Configuration
python-dotenv==1.0.0

# Testing (Optional)
pytest==7.4.3
httpx==0.25.2

# Logging
loguru==0.7.2

# Production Server
gunicorn==21.2.0
